Refactor to multi-module structure with Spring Boot 3.x and Java 17
Restructure the single-module project into three sub-modules following
the standard cicsdev multi-module layout:

  cics-java-liberty-springboot-jpa-app        — Spring Boot WAR
  cics-java-liberty-springboot-jpa-cicsbundle — CICS bundle plugin route
  cics-java-liberty-springboot-jpa-cicsbundle-eclipse — Eclipse export route

Spring Boot and Java version upgrades:
- Spring Boot 2.7 → 3.5.16 (Jakarta EE 10)
- Java 1.8 → 17 (Gradle toolchain / Maven release parameter)
- javax.persistence.* → jakarta.persistence.* in Emp.java
- web.xml upgraded to Servlet 6.0 / Jakarta EE 10 schema
- Hibernate dialect: DB2390Dialect → auto-detected (removed explicit
  property; Hibernate 6 infers dialect from JDBC metadata)

Build system alignment (Agents.md standards):
- Root build.gradle removed; root is now a pure Gradle settings holder
- settings.gradle: foojay toolchain resolver plugin, two submodule includes
- gradle.properties: java_version=17, auto-provisioning enabled
- Gradle wrapper: 7.5-rc-1 → 8.14.4
- Maven wrapper: 3.8.5 (with jar) → 3.9.12 (only-script, no jar)
- Root pom.xml: parent POM (packaging=pom) with Spring Boot BOM parent,
  two module declarations, maven-compiler-plugin with release parameter
- app/build.gradle: Spring Boot 3.5.16, Java toolchain, archiveFileName,
  bootWar disabled, eclipse.wtp block (prevents Gradle Refresh resetting
  WTP context-root and jst.web facet version)
- app/pom.xml: finalName set, spring-boot-starter-data-jpa dependency
- cicsbundle/build.gradle + pom.xml: cics-bundle-gradle/maven-plugin 1.0.8

Liberty and CICS configuration:
- etc/config/liberty/server.xml: servlet-6.0, jdbc-4.3 features;
  type-2 Db2 DataSource with type="javax.sql.DataSource" (required to
  avoid DSRA7015E — Liberty defaults to DB2XADataSource without this)
- src/main/webapp/WEB-INF/ibm-web-ext.xml: added with <web-ext> root
  element and correct context-root (Liberty silently ignores context-root
  if root element is wrong)

Eclipse metadata (out-of-box import experience):
- Root .project: both Gradle and Maven natures
- Root .classpath: minimal parent form (JavaSE-17 only)
- app .project / .classpath / .settings: full WTP project setup
- org.eclipse.wst.common.component: includes src/main/resources wb-resource
  entry so application.properties is included in Eclipse WTP export
- org.eclipse.wst.common.project.facet.core.xml: jst.web 6.0, jst.java 17
- cicsbundle and cicsbundle-eclipse Eclipse metadata added
- All modules: 2-line minimal org.eclipse.buildship.core.prefs

GitHub Actions:
- .github/workflows/build.yaml: 4-way matrix (gradle/gradlew/maven/mvnw)
  across Java 17/21/25 with copyright checker job
- .github/workflows/cleanup-old-runs.yml: weekly cleanup of old runs

Documentation and housekeeping:
- README.md: full rewrite — badges, Overview, Before You Start (Type-2
  DataSource config, CICS region setup, Liberty features), Downloading
  (Eclipse import steps), Building, Deploying (all 3 routes), Running,
  Troubleshooting, Contributing
- MAINTAINERS.md removed → CODEOWNERS added
- .gitignore updated to cicsdev standard
