+++ .github/workflows/build.yaml
@@ -1,0 +1,102 @@
+name: Build
+
+on:
+  push:
+    branches: ["main"]
+  pull_request:
+    branches: ["main"]
+permissions:
+  contents: read
+
+jobs:
+  check-copyright:
+    runs-on: ubuntu-latest
+    name: Check Copyright
+    permissions:
+      contents: write
+      pull-requests: write
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      - id: copyright-action
+        uses: cicsdev/.github/.github/actions/samples-copyright-checker@4134522d8109169bb8c460db841f94167ec2802f
+        with:
+          directory: './cics-java-liberty-springboot-jpa-app/'
+          file-extensions: '*.java'
+          base-copyright: 'Copyright IBM Corp. 2026'
+          token: ${{ secrets.GITHUB_TOKEN }}
+
+  build-maven:
+    name: Build Maven
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        jdk: [17, 21, 25]
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up JDK ${{ matrix.jdk }}
+        uses: actions/setup-java@v4
+        with:
+          java-version: ${{ matrix.jdk }}
+          distribution: "semeru"
+          cache: maven
+      - name: Build with Maven
+        run: mvn --batch-mode --update-snapshots --file pom.xml -Djava.version=${{ matrix.jdk }} verify
+
+  build-mvnw:
+    name: Build Maven Wrapper
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        jdk: [17, 21, 25]
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up JDK ${{ matrix.jdk }}
+        uses: actions/setup-java@v4
+        with:
+          java-version: ${{ matrix.jdk }}
+          distribution: "semeru"
+      - name: Build with Maven Wrapper
+        run: ./mvnw --batch-mode --update-snapshots --file pom.xml -Djava.version=${{ matrix.jdk }} verify
+
+  build-gradle:
+    name: Build Gradle
+    
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        jdk: [17, 21, 25]
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up JDK ${{ matrix.jdk }}
+      uses: actions/setup-java@v4
+      with:
+        java-version: ${{ matrix.jdk }}
+        distribution: 'semeru'
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+      with:
+        gradle-version: 8.14.4
+    - name: Build with Gradle
+      run: gradle clean build -Pjava_version=${{ matrix.jdk }}
+      
+  build-gradlew:
+    name: Build Gradle Wrapper
+    
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        jdk: [17, 21, 25]
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up JDK ${{ matrix.jdk }}
+      uses: actions/setup-java@v4
+      with:
+        java-version: ${{ matrix.jdk }}
+        distribution: 'semeru'
+    - name: Build with Gradle Wrapper
+      run: ./gradlew clean build -Pjava_version=${{ matrix.jdk }}
+
