+++ .settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,2 @@
-connection.project.dir=
 eclipse.preferences.version=1
+encoding/<project>=UTF-8
