--- .settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<faceted-project>
-	<fixed facet="jst.java"/>
-	<fixed facet="jst.web"/>
-	<installed facet="jst.web" version="2.4"/>
-	<installed facet="jst.java" version="1.8"/>
-</faceted-project>
