--- MAINTAINERS.md
@@ -1,7 +1,0 @@
-# Maintainers
-
-- Ivan D. Hargreaves [@IvanHargreaves](https://github.com/IvanHargreaves)
-- Alex Brown [@AlexBrown](https://github.com/SoftlySplinter)
-- Phil Wakelin [@PhilWakelin](https://github.com/PhilWakelin)
-
-*Last reviewed:* November 2024
