+++ cics-java-liberty-springboot-jpa-app/.gitignore
@@ -1,0 +1,4 @@
+/.gradle/
+/target/
+/build/
+/bin/
