+++ cics-java-liberty-springboot-jpa-app/.settings/org.eclipse.buildship.core.prefs
@@ -1,2 +1,2 @@
-connection.project.dir=
+connection.project.dir=..
 eclipse.preferences.version=1
