+++ cics-java-liberty-springboot-jpa-app/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+	<fixed facet="jst.java"/>
+	<fixed facet="jst.web"/>
+	<installed facet="jst.web" version="6.0"/>
+	<installed facet="jst.java" version="17"/>
+</faceted-project>
