+++ cics-java-liberty-springboot-jpa-app/build.properties
@@ -1,0 +1,3 @@
+source.. = src/main/java/
+bin.includes = META-INF/,.
+output.. = bin/
