+++ cics-java-liberty-springboot-jpa-app/pom.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  
+  <!-- Inherit from parent POM -->
+  <parent>
+    <groupId>com.ibm.cicsdev</groupId>
+    <artifactId>cics-java-liberty-springboot-jpa</artifactId>
+    <version>1.0.0</version>
+    <relativePath>../pom.xml</relativePath>
+  </parent>
+  
+  <!--  Application module properties -->
+  <artifactId>cics-java-liberty-springboot-jpa-app</artifactId>
+  <name>com.ibm.cicsdev.springboot.jpa.app</name>
+  <description>Demo project for Spring Boot JPA with CICS Liberty - Application</description>
+  
+  <properties>
+    <java.version>17</java.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <spring-boot.repackage.skip>true</spring-boot.repackage.skip>
+  </properties>
+  
+  <dependencies>   
+    <!-- Spring Boot web support -->
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-web</artifactId>
+    </dependency>
+
+    <!-- Compile against, but don't include TomCat in the runtime build -->
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-tomcat</artifactId>
+      <scope>provided</scope>
+    </dependency>
+
+    <!-- Spring JPA Support -->
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-data-jpa</artifactId>
+    </dependency>
+  </dependencies>
+
+  <!-- Package as an executable war (default jar) -->
+  <packaging>war</packaging>
+  <build>
+    <finalName>cics-java-liberty-springboot-jpa</finalName>
+    <plugins>
+      <plugin>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-maven-plugin</artifactId>
+      </plugin>
+    </plugins>
+  </build>
+
+</project>
