+++ cics-java-liberty-springboot-jpa-app/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.jpa.show-sql=true
+spring.data.jpa.repositories.bootstrap-mode=default
