+++ cics-java-liberty-springboot-jpa-app/src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<web-app xmlns="https://jakarta.ee/xml/ns/jakartaee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/web-app_6_0.xsd"
+         version="6.0">
+  <display-name>cics-java-liberty-springboot-jpa</display-name>
+  <login-config>
+    <auth-method>BASIC</auth-method>
+  </login-config>
+  <security-constraint>
+    <display-name>cics-java-liberty-springboot-jpa</display-name>
+    <web-resource-collection>
+      <web-resource-name>cics-java-liberty-springboot-jpa</web-resource-name>
+      <description>Protection rules for all servlets</description>
+      <url-pattern>/*</url-pattern>
+    </web-resource-collection>
+    <auth-constraint>
+      <description>All Authenticated users</description>
+      <role-name>cicsAllAuthenticated</role-name>
+    </auth-constraint>
+  </security-constraint>
+  <security-role>
+    <description>The CICS cicsAllAuthenticated role</description>
+    <role-name>cicsAllAuthenticated</role-name>
+  </security-role>
+</web-app>
