+++ cics-java-liberty-springboot-jpa-cicsbundle-eclipse/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,4 @@
+eclipse.preferences.version=1
+encoding//META-INF/cics.xml=UTF-8
+encoding/<project>=UTF-8
+encoding/cics-java-liberty-springboot-jpa.warbundle=UTF-8
