+++ cics-java-liberty-springboot-jpa-cicsbundle-eclipse/META-INF/cics.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<manifest xmlns="http://www.ibm.com/xmlns/prod/cics/bundle" id="cics-java-liberty-springboot-jpa-bundle" bundleMajorVer="1" bundleMinorVer="0" bundleMicroVer="0" bundleVersion="1" bundleRelease="0">
+    <meta_directives>
+        <timestamp>2026-01-01T00:00:00.000Z</timestamp>
+    </meta_directives>
+    <define name="cics-java-liberty-springboot-jpa-app" type="http://www.ibm.com/xmlns/prod/cics/bundle/WARBUNDLE" path="cics-java-liberty-springboot-jpa.warbundle"/>
+</manifest>
