+++ cics-java-liberty-springboot-jpa-cicsbundle-eclipse/cics-java-liberty-springboot-jpa.warbundle
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<warbundle symbolicname="cics-java-liberty-springboot-jpa-app" jvmserver="DFHWLP"/>
