+++ cics-java-liberty-springboot-jpa-cicsbundle/pom.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <!-- ================================================================ -->
+    <!-- Parent Project                                                   -->
+    <!-- ================================================================ -->
+    <parent>
+        <groupId>com.ibm.cicsdev</groupId>
+        <artifactId>cics-java-liberty-springboot-jpa</artifactId>
+        <version>1.0.0</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <!-- ================================================================ -->
+    <!-- Project Coordinates                                              -->
+    <!-- ================================================================ -->
+    <artifactId>cics-java-liberty-springboot-jpa-cicsbundle</artifactId>
+    <packaging>cics-bundle</packaging>
+    <name>CICS JPA Springboot - CICS Bundle</name>
+
+    <!-- ================================================================ -->
+    <!-- Dependencies                                                     -->
+    <!-- ================================================================ -->
+    <dependencies>
+        <!-- Application WAR -->
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>cics-java-liberty-springboot-jpa-app</artifactId>
+            <version>${project.version}</version>
+            <type>war</type>
+        </dependency>
+    </dependencies>
+
+    <!-- ================================================================ -->
+    <!-- Build Configuration                                              -->
+    <!-- ================================================================ -->
+    <build>
+        <plugins>
+            <!-- CICS Bundle Maven Plugin -->
+            <plugin>
+                <groupId>com.ibm.cics</groupId>
+                <artifactId>cics-bundle-maven-plugin</artifactId>
+                <version>1.0.8</version>
+                <extensions>true</extensions>
+                <configuration>
+                    <defaultjvmserver>${cics.jvmserver}</defaultjvmserver>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
