+++ etc/config/liberty/server.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<server description="Liberty server for CICS Spring Boot JPA sample">
+
+    <featureManager>
+        <!-- Required for Spring Boot 3.x (Jakarta EE 10) -->
+        <feature>servlet-6.0</feature>
+        <!-- Required for JPA/JDBC support -->
+        <feature>jdbc-4.3</feature>
+    </featureManager>
+
+    <!-- JDBC DataSource for Db2 via CICS type-2 connection.
+         type="javax.sql.DataSource" forces Liberty to use the non-XA DataSource interface.
+         Without this, Liberty defaults to DB2XADataSource which is not supported with driverType="2".
+         transactional="false" disables Liberty connection pooling transaction enlistment. -->
+    <dataSource id="t2" jndiName="jdbc/jpaDataSource" transactional="false" type="javax.sql.DataSource">
+        <jdbcDriver>
+            <library name="DB2LIB">
+                <fileset dir="/usr/lpp/db2v12/jdbc/classes" includes="db2jcc4.jar db2jcc_license_cisuz.jar"/>
+                <fileset dir="/usr/lpp/db2v12/jdbc/lib" includes="libdb2jcct2zos4_64.so"/>
+            </library>
+        </jdbcDriver>
+        <properties.db2.jcc currentSchema="DSN81110" driverType="2"/>
+        <connectionManager agedTimeout="0"/>
+    </dataSource>
+    
+</server>
