+++ gradle.properties
@@ -1,16 +1,14 @@
-#
-# Values in this file provide defaults to variables used in the 
-# gradle files.
-#
+java_version = 17
 
-# Normally we don't publish any built artifacts to a repository.
-# But if we do, these are the default values we use to indicate
-# where the files should be placed.
-#
-# These can be over-ridden from the command line 
-# with -Ppublish_repo_releases_url="file://my-folder" for example.
-# 
-# These values only have any effect if the publish goal is used.
-# For example: gradle build publish.
-publish_repo_releases_url = 'default-value-for-publish_repo_releases_url'
-publish_repo_releases_name = 'default-value-for-publish_repo_releases_name'
+# Gradle daemon improves build performance
+org.gradle.daemon=true
+
+# Enable parallel builds for faster compilation
+org.gradle.parallel=true
+
+# Enable toolchain auto-provisioning to download JDKs if not found locally
+org.gradle.java.installations.auto-detect=true
+org.gradle.java.installations.auto-download=true
+
+# Configuration cache disabled - CICS Bundle Plugin 1.0.8 not compatible
+# org.gradle.configuration-cache=true
