+++ pom.xml
@@ -1,78 +1,53 @@
-<project 
-  xmlns="http://maven.apache.org/POM/4.0.0" 
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-
-  <!-- Inherit defaults from Spring Boot -->
+  
+  <!-- Inherit defaults from Spring Boot Parent-->
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.7.0</version>
+    <version>3.5.16</version>
     <relativePath/>
   </parent>
-
-  <!-- Application properties -->
-  <groupId>com.ibm.cicsdev.springboot</groupId>
-  <artifactId>cics-java-liberty-springboot-jpa</artifactId>
-  <version>0.1.0</version>   
-  <name>cics-java-liberty-springboot.jpa</name>
-  <description>Demo JPA project for Spring Boot</description>
-  <properties>
-    <java.version>1.8</java.version>
-  </properties>
-   
-  <dependencies>
-
-    <!-- Spring Boot web support -->   
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-web</artifactId>
-    </dependency> 
-
-    <!-- Spring Integration JPA Support  -->
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-data-jpa</artifactId>
-    </dependency>
-    
-    <!-- Don't include TomCat in the runtime build, but do put it in WEB-INF so it can be run standalone a well as embedded --> 
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-tomcat</artifactId>
-      <scope>provided</scope>
-    </dependency>  
-  </dependencies>
   
-  <!-- Package as an executable war (default is jar) -->
-  <packaging>war</packaging>
+  <!-- ================================================================ -->
+  <!-- Project Coordinates                                              -->
+  <!-- ================================================================ -->
+  <groupId>com.ibm.cicsdev</groupId>
+  <artifactId>cics-java-liberty-springboot-jpa</artifactId>
+  <version>1.0.0</version>
+  <description>Demo project for Spring Boot JPA with CICS Liberty</description>
+  <packaging>pom</packaging>
+  
+  <properties>
+    <java.version>17</java.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <cics.jvmserver>DFHWLP</cics.jvmserver>
+  </properties>
+  
+  <!-- ================================================================ -->
+  <!-- Modules                                                          -->
+  <!-- ================================================================ -->
+  <modules>
+    <module>cics-java-liberty-springboot-jpa-app</module>
+    <module>cics-java-liberty-springboot-jpa-cicsbundle</module>
+  </modules>
+
+  <!-- ================================================================ -->
+  <!-- Build Configuration                                              -->
+  <!-- ================================================================ -->
   <build>
     <plugins>
+      <!-- Maven Compiler Plugin with release parameter -->
       <plugin>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-maven-plugin</artifactId>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <configuration>
+          <release>${java.version}</release>
+        </configuration>
       </plugin>
-   </plugins>
+    </plugins>
   </build>
-  
-  <!-- 
-  Publishes artifacts to here if the deploy goal is used. 
 
-  The values here can be passed on the maven command line
-  using the -D flag syntax 
-  
-  for example:
-  -Dpublish_repo_snapshots_name=my-repo 
-  -->
-  <distributionManagement>
-    <snapshotRepository>
-      <id>${publish_repo_snapshots_name}</id>
-      <url>${publish_repo_snapshots_url}</url>
-    </snapshotRepository>
-    <repository>
-      <id>${publish_repo_releases_name}</id>
-      <url>${publish_repo_releases_url}</url>
-    </repository>
-  </distributionManagement>
-  
 </project>
