--- src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.DB2390Dialect
-spring.jpa.show-sql=true
-spring.data.jpa.repositories.bootstrap-mode=default
